Major bug fix update of Word doc
1. Dimensions in centimeters now match dimensions in inches.
2. Excess styles removed from QuickStyles and Styles.
3. Heading 4 and Heading 5 styles redefined to be logical continuations
of Heading 1-3.
4. added “Names may optionally be placed in a single centered row
instead of at the top of each column.” per Dourish.
5. Caption Style updated to Justified [not Centered]; caption on Figure
1 updated to include “All 1-line captions should be centered; justify
longer captions.”
6. Figure 2 includes recommendation to place large figures and captions
in text boxes. Really weird old Section Break removed from the middle
of the document. (You know when you move a figure in the old format and
suddenly it all goes to one column? That’s what I fixed.)
7. References information consolidated at the end of the document and
edited. Make explicit the optional-but-encouraged nature of DOIs and
URLs in references.
8. Final page section break and advice regarding even final columns
restored.